+++ Mocolamma/Views/ChatMessagesView.swift
@@ -116,7 +116,7 @@ struct ChatMessagesView: View {
                 }
             }
         }
-        .onDrop(of: [.fileURL, .image], delegate: AreaImageDropDelegate(items: .constant([]), isDraggingOver: .constant(false), executor: executor))
+        .onDrop(of: [.fileURL, .image, .text], delegate: AreaImageDropDelegate(items: .constant([]), isDraggingOver: .constant(false), executor: executor))
     }
 }